requirements.txt
adagio==0.2.6
appdirs==1.4.4
asttokens==3.0.0
cloudpickle==3.1.1
comm==0.2.3
contourpy==1.3.3
coreforecast==0.0.16
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
et_xmlfile==2.0.0
executing==2.2.1
fonttools==4.59.2
fs==2.4.16
fsspec==2025.9.0
fugue==0.9.1
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
llvmlite==0.44.0
matplotlib==3.10.6
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numba==0.61.2
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
parso==0.8.5
patsy==1.0.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
Pygments==2.19.2
pyparsing==3.2.4
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyzmq==27.1.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
statsforecast==2.0.2
statsmodels==0.14.5
threadpoolctl==3.6.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
triad==0.9.8
tzdata==2025.2
utilsforecast==0.2.12
wcwidth==0.2.13
